Remove min drag size, but handle taps properly.
The minimum blur size was put there so that if you just tapped on the canvas it would at least do something.

But instead, now it will start the dragging of the rectangle so you realize that you're supposed to select and area by clicking again somewhere.

This doesn't quite work for touch so it will emit a warning if the user didn't drag and select an area.